feat: add customizations to Layout
