+++ .changeset/nasty-toes-grin.md
@@ -1,0 +1,7 @@
+---
+"gboost-ui": minor
+"gboost": patch
+---
+
+- Require passing `user`, `signOut` to Layout component for added flexibility
+- Add `AccountMenu` and `AccountSidebar` to Layout component for greater customization
