+++ packages/gboost/_templates/repo/create/ui/src/App.tsx.t
@@ -34,12 +34,14 @@ export function App() {
             signUpAttributes={signUpAttributes}
             title={title}
           >
-            {() => (
+            {({ signOut, user }) => (
               <Layout
                 className={theme}
                 logoSrc="/favicon-32x32.png"
                 pages={pages}
+                signOut={signOut}
                 title={title}
+                user={user}
               />
             )}
           </Authenticator>
